requirements.txt
praw
feedparser
beautifulsoup4
html5lib
python-dotenv
